+++ src/user-actions/user-actions.module.ts
@@ -2,6 +2,8 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ArticlesModule } from '../articles/articles.module';
 import { UsersModule } from '../users/users.module';
+import { AdminOpensController } from './controllers/admin-opens.controller';
+import { AdminSavedArticleController } from './controllers/admin-saved-article.controller';
 import { PersonalLinkRedirectController } from './controllers/personal-link-redirect.controller';
 import { SaveFromEmailController } from './controllers/save-from-email.controller';
 import { SavedArticlesController } from './controllers/saved-articles.controller';
@@ -17,7 +19,13 @@ import { SavedArticleService } from './services/saved-article.service';
     ArticlesModule,
     UsersModule,
   ],
-  controllers: [SavedArticlesController, SaveFromEmailController, PersonalLinkRedirectController],
+  controllers: [
+    SavedArticlesController,
+    SaveFromEmailController,
+    PersonalLinkRedirectController,
+    AdminSavedArticleController,
+    AdminOpensController,
+  ],
   providers: [SavedArticleService, PersonalArticleLinkService, SaveLinkSignatureService],
   exports: [SavedArticleService, PersonalArticleLinkService, SaveLinkSignatureService],
 })